reconfigured package name to prevent google services error, edited network code from main thread to prevent error... must fix post request format
