+++ app/google-services.json
@@ -10,7 +10,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:764849138700:android:978407211752cfe0d65229",
         "android_client_info": {
-          "package_name": "edu.temple.cloudmessagingapp"
+          "package_name": "edu.temple.contacttracer"
         }
       },
       "oauth_client": [
